refactor(emoji-picker, datepicker, scroller): name refactoring
