+++ components/emoji_picker/DtEmojiPicker.vue
