+++ components/emoji_picker/emoji_picker.stories.js
@@ -1,6 +1,6 @@
 import { action } from '@storybook/addon-actions';
 import { createTemplateFromVueFile } from '@/common/storybook_utils';
-import DtEmojiPicker from './emoji_picker.vue';
+import DtEmojiPicker from './DtEmojiPicker.vue';
 import DtEmojiPickerDefaultTemplate from './emoji_picker_default.story.vue';
 import DtEmojiPickerWithPopoverTemplate from './emoji_picker_popover.story.vue';
