+++ components/emoji_picker/emoji_picker_default.story.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script setup>
-import DtEmojiPicker from './emoji_picker.vue';
+import DtEmojiPicker from './DtEmojiPicker.vue';
 import { ref } from 'vue';
 
 const skinTone = ref('Default');
