+++ components/scroller/scroller_default.story.vue
@@ -48,7 +48,7 @@
 </template>
 
 <script setup>
-import DtScroller from './scroller.vue';
+import DtScroller from './DtScroller.vue';
 import { nextTick, ref } from 'vue';
 
 const items = ref(Array.from({ length: 50 }, (_, i) => ({
