+++ components/scroller/scroller_dynamic.story.vue
@@ -48,7 +48,7 @@
 </template>
 
 <script setup>
-import DtScroller from './scroller.vue';
+import DtScroller from './DtScroller.vue';
 import { nextTick, ref } from 'vue';
 
 import defaultAvatar from './person.png';
