+++ index.js
@@ -40,6 +40,8 @@ export * from './components/root_layout';
 export * from './components/icon';
 export * from './components/stack';
 export * from './components/scroller';
+export * from './components/emoji_picker';
+export * from './components/datepicker';
 
 /// Recipes
 export * from './recipes/comboboxes/combobox_with_popover';
